Add files via upload
get_pages_sqlru.py. Returns a list of all pages from SQL.RU forums that contain topics of the given author and contain given search string. Unlike built-in SQL.RU search, It returns a separate line for every page, not for every topic. It its move convenient because a topic can contain hundreeds of pages and only one of them contains the information required.
The list is returned in CSV format and contains topic names and page URLs.
1. Run get_pages_sqlru.py
2. When prompted enter author name if necessary
3. When prompted enter search string if necessary